+++ backend/libraries/types/src/notifications.rs
@@ -150,7 +150,7 @@ pub struct UserNotificationEnvelope {
     pub notification_bytes: ByteBuf,
     #[serde(rename = "t")]
     pub timestamp: TimestampMillis,
-    #[serde(rename = "f")]
+    #[serde(default, rename = "f")]
     pub fcm_data: FcmData,
 }