+++ backend/canisters/user/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove `correlation_id` ([#8097](https://github.com/open-chat-labs/open-chat/pull/8097))
 - Remove bot API Keys ([#8131](https://github.com/open-chat-labs/open-chat/pull/8131))
 - Support bot making inline replies ([#8146](https://github.com/open-chat-labs/open-chat/pull/8146))
+- Return newly created channel Ids after creating community ([#8156](https://github.com/open-chat-labs/open-chat/pull/8156))
 
 ### Fixed